backend/requirements.txt
langgraph
langchain-pinecone
pinecone-client
langchain-google-genai  # <--- Added this
python-dotenv